requirements.txt
africastalking==1.2.4
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contextlib2==21.6.0
cryptography==39.0.1
Deprecated==1.2.13
Django==4.1.6
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.4.2
jwt==1.3.1
oauthlib==3.2.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
schema==0.7.5
sqlparse==0.4.3
urllib3==1.26.14
wrapt==1.14.1
